Revert "Refactor out href logic into hook"
This reverts commit 9bb57f776aa03a4cfa39210b6a8019bcc391c9a6.
