+++ src/components/MDXComponents/MDXLink.tsx
@@ -1,14 +1,48 @@
 import Link from 'next/link';
 import ExternalLink from '@/components/ExternalLink';
-import { useMDXLinkHref } from './utils/useMDXLinkHref';
+import { useCurrentPlatform } from '@/utils/useCurrentPlatform';
+import { usePathWithoutHash } from '@/utils/usePathWithoutHash';
 
-export const MDXLink = ({ href, children, hash }) => {
+export const MDXLink = ({
+  href: hrefFromProps,
+  children,
+  hash: hashFromProps
+}) => {
+  let href = hrefFromProps;
+  let hash = hashFromProps;
   const isInternal = href && (href.startsWith('/') || href.startsWith('#'));
+  const baseURI = usePathWithoutHash();
+  const platform = useCurrentPlatform();
 
-  const nextLinkHref = useMDXLinkHref(href, hash);
+  if (href.startsWith('/') && href.includes('#')) {
+    hash = href.slice(href.indexOf('#'));
+    href = href.slice(0, href.indexOf('#'));
+  }
+
+  // For setting up a link when passed only a hash link
+  if (!href.includes('/') && href.startsWith('#')) {
+    hash = href;
+    href = baseURI.replace(platform, '[platform]');
+  }
+
+  const decodedUri = decodeURI(href);
+
+  // Check to see if we need to add the "platform" query param
+  // We shouldn't add it, if a specific platform is linked in the baseURI
+  const query = decodedUri.includes('[platform]')
+    ? { query: { platform } }
+    : {};
 
   return isInternal ? (
-    <Link href={nextLinkHref}>{children}</Link>
+    <Link
+      href={{
+        pathname: decodedUri,
+        ...(platform && query),
+        hash: hash
+      }}
+    >
+      {children}
+    </Link>
   ) : (
     <ExternalLink href={href}>{children}</ExternalLink>
   );
