+++ src/components/MDXComponents/__tests__/MDXLink.test.tsx
@@ -15,14 +15,6 @@ const routerMock = {
 
 jest.mock('next/router', () => routerMock);
 
-const MockedNextLink = ({ children }) => {
-  return children;
-};
-
-jest.mock('next/link', () => {
-  return MockedNextLink;
-});
-
 jest.mock('next/link', () => {
   const MockedNextLink = ({ children, href }) => {
     const hrefProp =
